Refactor RichTextEditor and formUtils; introduce rteUtils for TinyMCE options management
