+++ ui/app/src/components/PreviewConcierge/PreviewConcierge.tsx
@@ -186,6 +186,7 @@ import StandardAction from '../../models/StandardAction';
 import { createComponentId, pickShowContentFormAction } from '../../utils/system';
 import { popDialog, pushDialog } from '../../state/actions/dialogStack';
 import { nanoid } from 'nanoid';
+import { getCurrentLocale } from '../../utils/i18n';
 
 const issueDescriptorRequest = (props: {
 	site: string;
@@ -650,7 +651,8 @@ export function PreviewConcierge(props: PropsWithChildren<{}>) {
 							authoringBase: upToDateRefs.current.authoringBase,
 							site: upToDateRefs.current.siteId,
 							editModePadding: upToDateRefs.current.editModePadding,
-							rteConfig: upToDateRefs.current.rteConfig ?? {}
+							rteConfig: upToDateRefs.current.rteConfig ?? {},
+							locale: getCurrentLocale(upToDateRefs.current.user.username)
 						})
 					);
 					dispatch(guestCheckIn(payload));
