+++ ui/guest/src/iceRegistry.ts
@@ -19,7 +19,7 @@ import { DEFAULT_RECORD_DATA } from './utils/util';
 import * as contentTypeUtils from '@craftercms/studio-ui/utils/contentType';
 import * as Model from '@craftercms/studio-ui/utils/model';
 import { ContentInstance } from '@craftercms/studio-ui/models/ContentInstance';
-import {
+import type {
 	ContentType,
 	ContentTypeField,
 	ValidationKeys,
